Create oleveler module
