+++ oleveler/__init__.py
@@ -1,0 +1,8 @@
+from .main import *
+
+from .side_notebook_functions import \
+    displayImage, \
+    setDarkModePlotting, \
+    writeRSessionInfo
+
+from time import sleep
