+++ oleveler/version.py
@@ -1,0 +1,13 @@
+olevelerVersion = 1.1
+infoText = f'''
+Oleveler - omics data analysis tools for jupyter notebook
+
+Version {olevelerVersion}
+---
+by Chao DU
+Institute of Biology Leiden, Leiden University, the Netherlands
+c.du@biology.leidenuniv.nl
+durand[dot]dc[at]hot[no space]mail.com
+---
+
+'''
